Fix code structure for frontend
