+++ src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import {
   BrowserRouter as Router, Route, Routes,
 } from 'react-router-dom';
-import Login from './Components/Login'; // Asegúrate de que este path sea correcto
-import Register from './Components/Register'; // Asegúrate de que este path sea correcto
+import Login from './Components/auth/SignInForm'; // Asegúrate de que este path sea correcto
+import Register from './Components/auth/SignUpForm'; // Asegúrate de que este path sea correcto
 import AddGamesForm from './Components/games/AddGamesForm';
 import Session from './Components/session/Session';
 import MainPage from './Components/games/MainPage';
@@ -21,7 +21,7 @@ function App() {
         <Route path="/register" element={<Register />} />
         {/* <Route element={<RequireAuth />}> */}
         <Route exact path="/addGame" element={<AddGamesForm />} />
-        <Route path="/mainPage" element={<MainPage />} />
+        <Route path="/games" element={<MainPage />} />
         <Route path="/details/:gameId" element={<GameDetails />} />
         <Route path="/games/delete" element={<DeleteGame />} />
         <Route path="/reservations" element={<Reservations />} />
