--- src/Components/AddGameForm.js
@@ -1,70 +1,0 @@
-import React, { useState } from 'react';
-
-function AddGameForm() {
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
-  const [imageUrl, setImageUrl] = useState('');
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-
-    try {
-      const response = await fetch('http://localhost:3000/games', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          game: {
-            title: title,
-            description: description,
-            image_url: imageUrl,
-          },
-        }),
-      });
-
-      if (response.ok) {
-        console.log('Game added successfully');
-        // Handle any additional logic or redirection after adding the game
-      } else {
-        console.error('Error adding game');
-      }
-    } catch (error) {
-      console.error('Error adding game:', error);
-    }
-  };
-
-  return (
-    <div>
-      <h2>Add Game</h2>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label>Title:</label>
-          <input
-            type="text"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-          />
-        </div>
-        <div>
-          <label>Description:</label>
-          <textarea
-            value={description}
-            onChange={(e) => setDescription(e.target.value)}
-          />
-        </div>
-        <div>
-          <label>Image URL:</label>
-          <input
-            type="text"
-            value={imageUrl}
-            onChange={(e) => setImageUrl(e.target.value)}
-          />
-        </div>
-        <button type="submit">Add Game</button>
-      </form>
-    </div>
-  );
-}
-
-export default AddGameForm;
